DNSFS_Client/requirements.txt
dnspython>=2.4.2
cryptography>=41.0.3
argon2-cffi>=21.3.0
tqdm>=4.0.0
